CLI/pyproject.toml
[tool.poetry]
name = "cli"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
prompt-toolkit = "^3.0.39"
grpcio-tools = "^1.59.2"
termcolor = "^2.3.0"
nest-asyncio = "^1.5.8"
toml = "^0.10.2"
psutil = "^5.9.6"
pyinstaller = "^6.2.0"
icecream = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project=true